Moved a function inside to its own component.
